Extractor: Refactor parsing.py to a package
