+++ src/priviblur_extractor/__init__.py
@@ -1,4 +1,4 @@
 from .api import TumblrAPI
-from .parsers import parse_timeline, parse_blog_timeline, parse_item, parse_post_list
+from .parse import parse_timeline, parse_blog_timeline, parse_post_list
 from .helpers import exceptions as priviblur_exceptions
 from . import models
