+++ src/priviblur_extractor/parse/parsers.py
@@ -1,6 +1,6 @@
 import datetime
 
-from . import helpers, models
+from .. import helpers, models
 
 logger = helpers.LOGGER.getChild("parse")
